requirements.txt
python-telegram-bot>=21.10,<23.0
python-dotenv>=1.0,<2.0
pymongo>=4.8,<5.0
dnspython>=2.7,<3.0
